* Ok, can be uint8 too, but uses c in format, not u. Seems only worldinfo_collision and creature_canmove which should be 'c', will double check other structures next time.
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@900 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/AreaTrigger.h
@@ -36,7 +36,7 @@ enum AreaTriggerType
 typedef struct AreaTrigger
 {
 	uint32 AreaTriggerID;
-	uint32 Type;
+	uint8 Type;
 	uint32 Mapid;
 	uint32 PendingScreen;
 	char * Name;
